# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PG / Hostel Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      responses:
        "200":
          description: Properties
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Property" }
    post:
      operationId: createProperty
      tags: [properties]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PropertyInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Property" }
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Property
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PropertyDetail" }
    patch:
      operationId: updateProperty
      tags: [properties]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PropertyInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Property" }
    delete:
      operationId: deleteProperty
      tags: [properties]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      parameters:
        - { name: propertyId, in: query, required: false, schema: { type: integer } }
      responses:
        "200":
          description: Rooms with bed occupancy
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RoomWithBeds" }
    post:
      operationId: createRoom
      tags: [rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoomInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Room" }
  /rooms/{id}:
    patch:
      operationId: updateRoom
      tags: [rooms]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoomInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Room" }
    delete:
      operationId: deleteRoom
      tags: [rooms]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
  /beds/{id}/assign:
    post:
      operationId: assignBed
      tags: [beds]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssignBedRequest" }
      responses:
        "200":
          description: Assigned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Bed" }
  /beds/{id}/unassign:
    post:
      operationId: unassignBed
      tags: [beds]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Unassigned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Bed" }
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      parameters:
        - { name: status, in: query, required: false, schema: { type: string, enum: [paid, pending, overdue] } }
        - { name: tenantId, in: query, required: false, schema: { type: integer } }
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PaymentWithTenant" }
    post:
      operationId: createPayment
      tags: [payments]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PaymentInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payment" }
  /payments/{id}/mark-paid:
    post:
      operationId: markPaymentPaid
      tags: [payments]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MarkPaidRequest" }
      responses:
        "200":
          description: Marked paid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payment" }
  /payments/generate-monthly:
    post:
      operationId: generateMonthlyRent
      tags: [payments]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GenerateMonthlyRequest" }
      responses:
        "200":
          description: Generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  generated: { type: integer }
  /complaints:
    get:
      operationId: listComplaints
      tags: [complaints]
      parameters:
        - { name: status, in: query, required: false, schema: { type: string, enum: [open, in_progress, resolved] } }
      responses:
        "200":
          description: Complaints
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ComplaintWithTenant" }
    post:
      operationId: createComplaint
      tags: [complaints]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ComplaintInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Complaint" }
  /complaints/{id}:
    patch:
      operationId: updateComplaintStatus
      tags: [complaints]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateComplaintStatusRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Complaint" }
components:
  schemas:
...
    AssignBedRequest:
      type: object
      required: [tenantId]
      properties:
        tenantId: { type: integer }
    MarkPaidRequest:
      type: object
      required: [method]
      properties:
        method: { type: string, enum: [cash, upi, bank_transfer, card] }
    GenerateMonthlyRequest:
      type: object
      required: [month, year]
      properties:
        month: { type: integer, minimum: 1, maximum: 12 }
        year: { type: integer }
    UpdateComplaintStatusRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [open, in_progress, resolved] }
        resolutionNote: { type: string, nullable: true }
...
    Property:
      type: object
      required: [id, name, address, city, totalFloors, contactPhone, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        address: { type: string }
        city: { type: string }
        totalFloors: { type: integer }
        contactPhone: { type: string }
        description: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    PropertyInput:
      type: object
      required: [name, address, city, totalFloors, contactPhone]
      properties:
        name: { type: string }
        address: { type: string }
        city: { type: string }
        totalFloors: { type: integer }
        contactPhone: { type: string }
        description: { type: string, nullable: true }
    PropertyDetail:
      allOf:
        - $ref: "#/components/schemas/Property"
        - type: object
          required: [totalRooms, totalBeds, occupiedBeds]
          properties:
            totalRooms: { type: integer }
            totalBeds: { type: integer }
            occupiedBeds: { type: integer }
    Room:
      type: object
      required: [id, propertyId, roomNumber, floor, capacity, monthlyRent, roomType]
      properties:
        id: { type: integer }
        propertyId: { type: integer }
        roomNumber: { type: string }
        floor: { type: integer }
        capacity: { type: integer }
        monthlyRent: { type: number }
        roomType: { type: string, enum: [single, double, triple, quad, dormitory] }
        amenities: { type: string, nullable: true }
    RoomInput:
      type: object
      required: [propertyId, roomNumber, floor, capacity, monthlyRent, roomType]
      properties:
        propertyId: { type: integer }
        roomNumber: { type: string }
        floor: { type: integer }
        capacity: { type: integer }
        monthlyRent: { type: number }
        roomType: { type: string, enum: [single, double, triple, quad, dormitory] }
        amenities: { type: string, nullable: true }
    Bed:
      type: object
      required: [id, roomId, bedLabel, isOccupied]
      properties:
        id: { type: integer }
        roomId: { type: integer }
        bedLabel: { type: string }
        isOccupied: { type: boolean }
        tenantId: { type: integer, nullable: true }
        tenantName: { type: string, nullable: true }
    RoomWithBeds:
      allOf:
        - $ref: "#/components/schemas/Room"
        - type: object
          required: [propertyName, beds, occupiedCount]
          properties:
            propertyName: { type: string }
            beds:
              type: array
              items: { $ref: "#/components/schemas/Bed" }
            occupiedCount: { type: integer }
...
    Payment:
      type: object
      required: [id, tenantId, amount, status, dueDate, month, year, createdAt]
      properties:
        id: { type: integer }
        tenantId: { type: integer }
        amount: { type: number }
        status: { type: string, enum: [paid, pending, overdue] }
        dueDate: { type: string, format: date }
        paidDate: { type: string, format: date, nullable: true }
        month: { type: integer }
        year: { type: integer }
        method: { type: string, enum: [cash, upi, bank_transfer, card], nullable: true }
        notes: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    PaymentInput:
      type: object
      required: [tenantId, amount, dueDate, month, year]
      properties:
        tenantId: { type: integer }
        amount: { type: number }
        dueDate: { type: string, format: date }
        month: { type: integer, minimum: 1, maximum: 12 }
        year: { type: integer }
        notes: { type: string, nullable: true }
    PaymentWithTenant:
      allOf:
        - $ref: "#/components/schemas/Payment"
        - type: object
          required: [tenantName]
          properties:
            tenantName: { type: string }
            roomNumber: { type: string, nullable: true }
            propertyName: { type: string, nullable: true }
    Complaint:
      type: object
      required: [id, tenantId, category, title, description, priority, status, createdAt]
      properties:
        id: { type: integer }
        tenantId: { type: integer }
        category: { type: string, enum: [plumbing, electricity, cleaning, internet, furniture, security, other] }
        title: { type: string }
        description: { type: string }
        priority: { type: string, enum: [low, medium, high] }
        status: { type: string, enum: [open, in_progress, resolved] }
        resolutionNote: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        resolvedAt: { type: string, format: date-time, nullable: true }
    ComplaintInput:
      type: object
      required: [tenantId, category, title, description, priority]
      properties:
        tenantId: { type: integer }
        category: { type: string, enum: [plumbing, electricity, cleaning, internet, furniture, security, other] }
        title: { type: string }
        description: { type: string }
        priority: { type: string, enum: [low, medium, high] }
    ComplaintWithTenant:
      allOf:
        - $ref: "#/components/schemas/Complaint"
        - type: object
          required: [tenantName]
          properties:
            tenantName: { type: string }
            roomNumber: { type: string, nullable: true }
            propertyName: { type: string, nullable: true }