# lib/api-zod/src/generated/types/complaintWithTenant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PG / Hostel Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Complaint } from "./complaint";

export type ComplaintWithTenant = Complaint & {
  tenantName: string;
  roomNumber?: string | null;
  propertyName?: string | null;
};