# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PG / Hostel Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Property {
  id: number;
  name: string;
  address: string;
  city: string;
  totalFloors: number;
  contactPhone: string;
  description?: string | null;
  createdAt: Date;
}