# lib/api-zod/src/generated/types/roomWithBeds.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PG / Hostel Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Bed } from "./bed";
import type { Room } from "./room";

export type RoomWithBeds = Room & {
  propertyName: string;
  beds: Bed[];
  occupiedCount: number;
};