# lib/api-zod/src/generated/types/updateComplaintStatusRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PG / Hostel Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateComplaintStatusRequestStatus } from "./updateComplaintStatusRequestStatus";

export interface UpdateComplaintStatusRequest {
  status: UpdateComplaintStatusRequestStatus;
  resolutionNote?: string | null;
}