# lib/db/src/schema/payments.ts
import { pgTable, serial, integer, numeric, text, date, timestamp } from "drizzle-orm/pg-core";
import { tenantsTable } from "./tenants";

export const paymentsTable = pgTable("payments", {
  id: serial("id").primaryKey(),
  tenantId: integer("tenant_id").notNull().references(() => tenantsTable.id, { onDelete: "cascade" }),
  amount: numeric("amount", { precision: 10, scale: 2 }).notNull(),
  status: text("status").notNull().default("pending"),
  dueDate: date("due_date").notNull(),
  paidDate: date("paid_date"),
  month: integer("month").notNull(),
  year: integer("year").notNull(),
  method: text("method"),
  notes: text("notes"),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
});

export type Payment = typeof paymentsTable.$inferSelect;
export type InsertPayment = typeof paymentsTable.$inferInsert;